Remove dock icon.
This kills the electron application as soon as it starts without any
delay, so the dock icon never appears.

Otherwise, running this multiple times would cause a flickering in the
dock icon.

This works by closing electron, which usually also kills the process,
but we hook into the `will-close` and prevent it from closing using
`deasync`.